examples/jsonapi/requirements.txt


marshmallow-jsonapi